lark-9: Made the exportables table into reusable service.
